Fix build warning on 32-bit int platfowms.  Bug 2678
(cherry picked from commit abca11df7f354c123b0ff8a9bfb89a669d7742d4)

+++ src/src/pdkim/pdkim.c
@@ -1324,7 +1324,8 @@ check_bare_ed25519_pubkey(pdkim_pubkey * p)
 int excess = p->key.len - 32;
 if (excess > 0)
   {
-  DEBUG(D_acl) debug_printf("DKIM: unexpected pubkey len %lu\n", p->key.len);
+  DEBUG(D_acl)
+    debug_printf("DKIM: unexpected pubkey len %lu\n", (unsigned long) p->key.len);
   p->key.data += excess; p->key.len = 32;
   }
 }
